sol-util: rename sol-util-linux to sol-util-file
Since they're file functions, not only linux might need them.
Also, we'll start to use sol-util-$(PLATFORM) in the next commit.
