+++ src/modules/flow/keyboard/keyboard.c
@@ -47,7 +47,7 @@
 #include "sol-mainloop.h"
 #include "sol-vector.h"
 #include "sol-util.h"
-#include "sol-util-linux.h"
+#include "sol-util-file.h"
 
 struct keyboard_common_data {
     struct sol_flow_node *node;
