+++ src/modules/flow/process/subprocess.c
@@ -34,7 +34,7 @@
 
 #include "sol-platform-linux.h"
 #include "sol-util.h"
-#include "sol-util-linux.h"
+#include "sol-util-file.h"
 
 #include <fcntl.h>
 #include <signal.h>
