+++ src/shared/sol-util-file.h
